+++ include/universal/maths.hpp
@@ -5,36 +5,40 @@
 
 // Add comments in tmr or weekend to explain the maths and don't forget to write notes on the formulas used and the correct notation
 
+// Universal maths namespace
 namespace maths {
-	static constexpr double pi = 3.14159265358979323846;
-	static constexpr double e = 2.71828182845904523536;
+	static constexpr float pi = 3.14159265358979323846; // Pi constant
+	static constexpr float e = 2.71828182845904523536; // Euler's number
 
 
 	// 3D Vector structure
 	struct Vector3 {
-		// Initial data
-		double x, y, z;
+		// Initial data members
+		float x, y, z;
 
-		// Constructor. 
-		Vector3(double _x, double _y, double _z) : x(_x), y(_y), z(_z) {};
+		// Constructor. _x means the constructor parameter to differentiate from the member variable
+		Vector3(float _x, float _y, float _z) : x(_x), y(_y), z(_z) {};
 
-		double dot(const Vector3& other) const {
+		// Dot product method
+		float dot(const Vector3& other) const {
 			return (x * other.x) + (y * other.y) + (z * other.z);
 		}
 
+		// Addition operator overload
 		Vector3 operator+(const Vector3& other) const { return Vector3(x + other.x, y + other.y, z + other.z); }
 
 
-		double magnitude() {
+		// Magnitude method
+		float magnitude() {
 			return std::sqrt((x * x) + (y * y) + (z * z));
 		};
 	};
 
-	static constexpr double degreesToRadians(double degrees) {
+	static constexpr float degreesToRadians(float degrees) {
 		return degrees * (pi / 180.0);
 	}
 
-	static constexpr double radiansToDegrees(double radians) {
+	static constexpr float radiansToDegrees(float radians) {
 		return radians * (180.0 / pi);
 	}
