+++ include/universal/physics.hpp
@@ -1,3 +1,3 @@
 namespace physics {
-	constexpr double gravity = 9.80665; // m/s^2
+	constexpr float gravity = 9.80665; // m/s^2
 }
